BOJ_9019_DSLRCalcul_BF,BFS
BOJ_9019_DSLRCalcul_BF,BFS
Simple is the best.
When I use int to string and calculate.
It's too many process. So, it's slow.
And I used vector in struct.
But, It didn' work well as I think.
So, I should change the way.
I use string and add the character.